Moved access token logic out of read activities model and into controller
